+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## HEAD
 
 - BREAKING CHANGE: Remove `ping` and `pong` signals
+- BREAKING CHANGE: Remove `username_list_received` signal
 - Update message handler
 
 ## 2023-09-28 - v0.2.0